--- src/chatserver/chain/__init__.py
@@ -1,1 +1,0 @@
-from .lightning_client import LightningChain
